+++ CHANGELOG.md
@@ -11,6 +11,10 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 - dipdup: Python 3.12 is now required to run DipDup.
 - dipdup: Pydantic v2 migration.
 
+### Removed
+
+- evm.subsquid: `node_only` index config option has been removed; use `evm.node` as datasource instead.
+
 ## [7.5.4] - 2024-04-09
 
 ### Fixed